Move kernel factory code to separate module
